pyproject.toml
[project]
name = "mecagent-technical-test"
version = "0.1.0"
description = "Just a technical test"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.8.0",
    "cadquery>=2.5.2",
    "datasets>=3.6.0",
    "ipykernel>=6.29.5",
    "loguru>=0.7.3",
    "numpy>=2.3.0",
    "pillow>=11.2.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.0",
    "scipy>=1.15.3",
    "torch>=2.7.1",
    "transformers>=4.52.4",
    "trimesh>=4.6.11",
]
